Fix wedge in wide to narrow write module
The bit width for the per-flit size used the smaller bit width,
meaning it was always zero. This caused the module to think
it had wrapped all the way around to zero, and so deq the
aw flit, and wedge.

The other components look like they might be okay at first
glance, but may need further investigation/testing.